fix(help-find): avoid cloning meta when preparing join query
Applying the base query’s `meta` object by reference here should
resolve an issue where prototype information of our
`meta.leasedConnection` object is lost due to the deep clone.

When the leased connection object loses its original prototype, we’re
liable to see breakage in adapters due to expected properties no longer
being present. For example, sails-mongo requires an accessible
`collection` property, though this will be lost in our clone and thus
yield an error when an attempt is made to call it as a function.

For the purpose of performance (perhaps a micro-optimisation for some),
this commit has our `meta` object shared between the parent and child
queries. This appears to be safe in my testing with sails-mongo, but
could cause issues if an adapter decided to modify the `meta` object.